requirements.txt
Flask
gunicorn
google-cloud-storage
pandas
scikit-learn

